+++ src/pages/Home.js
@@ -1,24 +1,14 @@
-import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
 import Popup from 'reactjs-popup';
 import Todo from '../js/Todo';
 import { IoMdClose } from 'react-icons/io';
 import 'reactjs-popup/dist/index.css';
 import DisplayJournal from '../js/displayEntries.js';
+import GenerateMessage from '../js/botPrompt.js';
 
 const Home = () => {
-  const [randomMessage, setRandomMessage] = useState('Talk to me! Click me for a surprise.');
-  const messages = [
-    'You are a shining star!',
-    'Keep up the great work!',
-    'Believe in yourself!',
-    'You can achieve anything!',
-    'Stay positive and happy!'
-  ];
-
-  const handleClick = () => {
-    const randomMsg = messages[Math.floor(Math.random() * messages.length)];
-    setRandomMessage(randomMsg);
+  const displayMascot = () => {
+    return <GenerateMessage />;
   };
 
   return (
@@ -51,18 +41,7 @@ const Home = () => {
       <div>
         <DisplayJournal />
       </div>
-      <div className="interactive-container">
-        <button onClick={handleClick} style={{ border: 'none', background: 'none', padding: '0' }}>
-          <img
-            src="public/starfish_img.png"
-            alt="Starfish"
-            className="button"
-            style={{ cursor: 'pointer' }}
-          />
-        </button>
-      </div>
-      <footer>This is a temp footer</footer>
-      <div className="textbox">{randomMessage}</div>
+      {displayMascot()}
     </main>
   );
 };
